+++ src/pyday/Data2PDF/BasicStyle.py
@@ -27,27 +27,34 @@ Text = [
 # TA_RIGHT = 2
 # TA_JUSTIFY = 4
 
+# A4.w = 468
+
 Image = [
         [
             
         ]
     ]
+
 Table = [
-        [
+        [   
+            {"colWidths":None, "rowHeights":None, "splitByRow":1, "repeatRows":22, "repeatCols":0, "rowSplitRange":None, "spaceBefore":12, "spaceAfter":12, "cornerRadii":None},
             ("ALIGN", (0, 0), (-1, -1), "CENTER"),  # 置中對齊
             ("FONTNAME", (0, 0), (-1, -1), "Noto_Sans_TC_Regular"),  # 字體
             ("VALIGN", (0, 0), (-1, -1), "MIDDLE"),  # 上下置中
             ("GRID", (0, 0), (-1, -1), 0.5, HexColor(0x000000) ),  # 框線黑色，寬度0.5
         ],
         [
-                ("BACKGROUND", (0, 0), (-1, -1), HexColor("#F2B705")),  # 表头背景颜色
-                ("ALIGN", (0, 0), (-1, -1), "LEFT"),  # 向左對齊
-                ("FONTNAME", (0, 0), (-1, -1), "Noto_Sans_TC_Regular"),  # 字體
-                ("FONTSIZE", (0, 0), (-1, -1), 12),  # 表头字体大小
-                ("VALIGN", (0, 0), (-1, -1), "MIDDLE"),  # 上下置中
-                ("GRID", (0, 0), (-1, -1), 0.5, HexColor(0x000000)),  # 框線黑色，寬度0.5
+            {"colWidths":None, "rowHeights":None, "splitByRow":1, "repeatRows":0, "repeatCols":0, "rowSplitRange":None, "spaceBefore":None, "spaceAfter":None, "cornerRadii":None},
+            ("BACKGROUND", (0, 0), (-1, -1), HexColor("#F2B705")),  # 表头背景颜色
+            ("ALIGN", (0, 0), (-1, -1), "LEFT"),  # 向左對齊
+            ("FONTNAME", (0, 0), (-1, -1), "Noto_Sans_TC_Regular"),  # 字體
+            ("FONTSIZE", (0, 0), (-1, -1), 12),  # 表头字体大小
+            ("VALIGN", (0, 0), (-1, -1), "MIDDLE"),  # 上下置中
+            ("GRID", (0, 0), (-1, -1), 0.5, HexColor(0x000000)),  # 框線黑色，寬度0.5
         ]
     ]
+
+# Table(data, colWidths=None, rowHeights=None, style=None, splitByRow=1, repeatRows=0, repeatCols=0, rowSplitRange=None, spaceBefore=None, spaceAfter=None, cornerRadii=None)
 
 # Table(
 #     data[1],
@@ -56,7 +63,7 @@ Table = [
 #     rowHeights=22,
 #     spaceBefore=12,
 #     # spaceAfter=12,
-# )
+# ) 
 
 # Table(
 #     data[1],
@@ -65,3 +72,5 @@ Table = [
 #     spaceBefore=12,
 #     spaceAfter=12,
 # )
+
+# https://docs.reportlab.com/reportlab/userguide/ch7_tables/#tabledata-colwidthsnone-rowheightsnone-stylenone-splitbyrow1-repeatrows0-repeatcols0-rowsplitrangenone-spacebeforenone-spaceafternone-cornerradiinone