+++ src/pyday/__init__.py
@@ -4,3 +4,5 @@ from .DataVis.DataVis import *
 
 from .Tool.dirTree import *
 from .Tool.ChangLanguage import *
+
+__version__ = '0.0.23'