+++ src/test.py
@@ -5,5 +5,6 @@
 
 from pyday import Data2PDF
 d2pdf = Data2PDF()
-d2pdf.inFile("test2.json")
-d2pdf.toFile("test2.pdf")
+d2pdf.inFile("test.json")
+print( d2pdf.getTableStyle() )
+# d2pdf.toFile("test2.pdf")