+++ doc/changelog.md
@@ -1,4 +1,9 @@
 # Change Log(更新日志)
+## pyday 2023-09-08(v0.0.24)
+- [AnsonCar] 整埋文件夾
+- [AnsonCar] 重整Tool 
+- [AnsonCar] 新增Tool ChangeLanguage inFile 和 toFile 功能
+
 ## pyday 2023-08-09(v0.0.23)
 - [新增] [AnsonCar] Tool dirTree 顯示文件樹狀結構工具
 - [重整] [AnsonCar] DataReader 功能